update xpu graph trees support

+++ torch/csrc/Module.cpp
@@ -3212,7 +3212,7 @@ Call this whenever a new thread is created in order to propagate values from
 
   py_module.def(
       "_construct_storage_from_data_pointer",
-      [](int64_t data_ptr, c10::Device device, size_t size_bytes) {
+      [](uintptr_t data_ptr, c10::Device device, size_t size_bytes) {
         auto external_data_ptr = at::DataPtr(
             // NOLINTNEXTLINE(performance-no-int-to-ptr)
             reinterpret_cast<void*>(data_ptr),
